Nimra/Sprint2/Nimra_Sprint2_CICD/requirements-aws.txt
aws-cdk.core==1.135.0
aws-cdk.assertions==1.135.0
aws-cdk.pipelines==1.135.0
aws_cdk.aws_codepipeline_actions==1.135.0
boto3
pytest